fix: data race running many processes concurrently (#167)
* Add failing test

This test is flaky rather than always failing, but with 1000 runs it's sufficiently flaky to be useful.

* Extract file reading in safer mechanism

There are lots of potential races when reading from process file handles. This approach extracts the handling into the file handle itself, which makes it easier to get the lifecycle right.

This also unifies reading from the file handles into one place, as we were losing data between the readabilityHandler and sync readToEnd calls.

* Fix linting issues

* Migrate the test to Swift Testing

---------

Co-authored-by: Pedro <pedro@pepicrft.me>